requirements.txt
fastapi[all]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
azure-storage-blob
pandas
